Swap scratch X and Y use
We use a lot of scratch for CGIA and it should be in scratch X,
which is shared with Core 1 stack. Core 1 renderer uses little stack.
Core 0 runs general purpose firmware code and can use
a lot of stack, so we can use only a little of scratch Y,
which is shared with Core 0 stack.